+++ bundles/org.openhab.binding.easee/src/main/java/org/openhab/binding/easee/internal/handler/EaseeChargerHandler.java
@@ -21,8 +21,6 @@ import java.util.concurrent.atomic.AtomicReference;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.binding.easee.internal.AtomicReferenceTrait;
-import org.openhab.binding.easee.internal.EaseeBindingConstants;
 import org.openhab.binding.easee.internal.Utils;
 import org.openhab.binding.easee.internal.command.EaseeCommand;
 import org.openhab.binding.easee.internal.command.charger.ChangeConfiguration;
@@ -32,18 +30,12 @@ import org.openhab.binding.easee.internal.command.charger.LatestChargingSession;
 import org.openhab.binding.easee.internal.command.charger.SendCommand;
 import org.openhab.binding.easee.internal.command.charger.SendCommandPauseResume;
 import org.openhab.binding.easee.internal.command.charger.SendCommandStartStop;
-import org.openhab.binding.easee.internal.config.EaseeConfiguration;
 import org.openhab.binding.easee.internal.connector.CommunicationStatus;
-import org.openhab.core.thing.Bridge;
 import org.openhab.core.thing.Channel;
 import org.openhab.core.thing.Thing;
 import org.openhab.core.thing.ThingStatus;
 import org.openhab.core.thing.ThingStatusDetail;
-import org.openhab.core.thing.ThingStatusInfo;
-import org.openhab.core.thing.binding.BaseThingHandler;
 import org.openhab.core.types.Command;
-import org.openhab.core.types.State;
-import org.openhab.core.types.UnDefType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -56,7 +48,7 @@ import com.google.gson.JsonObject;
  * @author Alexander Friese - initial contribution
  */
 @NonNullByDefault
-public class EaseeChargerHandler extends BaseThingHandler implements EaseeThingHandler, AtomicReferenceTrait {
+public class EaseeChargerHandler extends EaseeBaseThingHandler {
     private final Logger logger = LoggerFactory.getLogger(EaseeChargerHandler.class);
 
     /**
@@ -72,25 +64,6 @@ public class EaseeChargerHandler extends BaseThingHandler implements EaseeThingH
     }
 
     @Override
-    public void bridgeStatusChanged(ThingStatusInfo bridgeStatusInfo) {
-        super.bridgeStatusChanged(bridgeStatusInfo);
-        if (bridgeStatusInfo.getStatus() == ThingStatus.ONLINE) {
-            logger.debug("bridgeStatusChanged: ONLINE");
-            if (isInitialized()) {
-                startPolling();
-            }
-        } else {
-            logger.debug("bridgeStatusChanged: NOT ONLINE");
-            if (isInitialized()) {
-                if (bridgeStatusInfo.getStatus() == ThingStatus.UNKNOWN) {
-                    updateStatus(ThingStatus.UNKNOWN, ThingStatusDetail.NONE, STATUS_WAITING_FOR_BRIDGE);
-                }
-                stopPolling();
-            }
-        }
-    }
-
-    @Override
     public void initialize() {
         logger.debug("About to initialize Charger");
         logger.debug("Easee Charger initialized with id: {}", getId());
@@ -99,10 +72,6 @@ public class EaseeChargerHandler extends BaseThingHandler implements EaseeThingH
         startPolling();
 
         enqueueCommand(new Charger(this, getId(), this::updatePropertiesAndOnlineStatus));
-    }
-
-    public String getId() {
-        return getConfig().get(EaseeBindingConstants.THING_CONFIG_ID).toString();
     }
 
     private void updatePropertiesAndOnlineStatus(CommunicationStatus status, JsonObject charger) {
@@ -136,7 +105,8 @@ public class EaseeChargerHandler extends BaseThingHandler implements EaseeThingH
     /**
      * Start the polling.
      */
-    private void startPolling() {
+    @Override
+    protected void startPolling() {
         updateJobReference(dataPollingJobReference, scheduler.scheduleWithFixedDelay(this::pollingRun,
                 POLLING_INITIAL_DELAY, getBridgeConfiguration().getDataPollingInterval(), TimeUnit.SECONDS));
 
@@ -147,7 +117,8 @@ public class EaseeChargerHandler extends BaseThingHandler implements EaseeThingH
     /**
      * Stops the polling.
      */
-    private void stopPolling() {
+    @Override
+    protected void stopPolling() {
         cancelJobReference(dataPollingJobReference);
         cancelJobReference(sessionDataPollingJobReference);
     }
@@ -156,7 +127,7 @@ public class EaseeChargerHandler extends BaseThingHandler implements EaseeThingH
      * Poll the Easee Cloud API one time.
      */
     void pollingRun() {
-        String chargerId = getConfig().get(EaseeBindingConstants.THING_CONFIG_ID).toString();
+        String chargerId = getId();
         logger.debug("polling charger data for {}", chargerId);
 
         // proceed if charger is online
@@ -169,7 +140,7 @@ public class EaseeChargerHandler extends BaseThingHandler implements EaseeThingH
      * Poll the Easee Cloud API session data endpoint one time.
      */
     void sessionDataPollingRun() {
-        String chargerId = getConfig().get(EaseeBindingConstants.THING_CONFIG_ID).toString();
+        String chargerId = getId();
         logger.debug("polling session data for {}", chargerId);
 
         // proceed if charger is online
@@ -180,7 +151,7 @@ public class EaseeChargerHandler extends BaseThingHandler implements EaseeThingH
 
     /**
      * updates online status depending on online information received from the API. this is called by the SiteState
-     * Command which retrieves whole site data inclusing charger status.
+     * Command which retrieves whole site data inclusing charger online status.
      *
      */
     public void setOnline(boolean isOnline) {
@@ -191,87 +162,9 @@ public class EaseeChargerHandler extends BaseThingHandler implements EaseeThingH
         }
     }
 
-    /**
-     * result processor to handle online status updates
-     *
-     * @param status of command execution
-     * @param jsonObject json respone result
-     */
-    protected final void updateOnlineStatus(CommunicationStatus status, JsonObject jsonObject) {
-        String msg = Utils.getAsString(jsonObject, JSON_KEY_ERROR_TITLE);
-        if (msg == null || msg.isBlank()) {
-            msg = status.getMessage();
-        }
-
-        switch (status.getHttpCode()) {
-            case OK:
-            case ACCEPTED:
-                super.updateStatus(ThingStatus.ONLINE, ThingStatusDetail.NONE);
-                break;
-            default:
-                super.updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.CONFIGURATION_ERROR, msg);
-        }
-    }
-
-    /**
-     * Disposes the thing.
-     */
-    @Override
-    public void dispose() {
-        logger.debug("Handler disposed.");
-        stopPolling();
-    }
-
-    /**
-     * will update all channels provided in the map
-     */
-    @Override
-    public void updateChannelStatus(Map<Channel, State> values) {
-        logger.debug("Handling charger channel update.");
-
-        for (Channel channel : values.keySet()) {
-            if (getThing().getChannels().contains(channel)) {
-                State value = values.get(channel);
-                if (value != null) {
-                    logger.debug("Channel is to be updated: {}: {}", channel.getUID().getAsString(), value);
-                    updateState(channel.getUID(), value);
-                } else {
-                    logger.debug("Value is null or not provided by Easee Cloud (channel: {})",
-                            channel.getUID().getAsString());
-                    updateState(channel.getUID(), UnDefType.UNDEF);
-                }
-            } else {
-                logger.debug("Could not identify channel: {} for model {}", channel.getUID().getAsString(),
-                        getThing().getThingTypeUID().getAsString());
-            }
-        }
-    }
-
-    @Override
-    public void enqueueCommand(EaseeCommand command) {
-        EaseeBridgeHandler bridgeHandler = getBridgeHandler();
-        if (bridgeHandler != null) {
-            bridgeHandler.enqueueCommand(command);
-        } else {
-            // this should not happen
-            logger.warn("no bridge handler found");
-        }
-    }
-
-    private @Nullable EaseeBridgeHandler getBridgeHandler() {
-        Bridge bridge = getBridge();
-        return bridge == null ? null : ((EaseeBridgeHandler) bridge.getHandler());
-    }
-
-    @Override
-    public EaseeConfiguration getBridgeConfiguration() {
-        EaseeBridgeHandler bridgeHandler = getBridgeHandler();
-        return bridgeHandler == null ? new EaseeConfiguration() : bridgeHandler.getBridgeConfiguration();
-    }
-
     @Override
     public EaseeCommand buildEaseeCommand(Command command, Channel channel) {
-        String chargerId = getConfig().get(EaseeBindingConstants.THING_CONFIG_ID).toString();
+        String chargerId = getId();
 
         switch (Utils.getWriteCommand(channel)) {
             case COMMAND_CHANGE_CONFIGURATION:
@@ -289,10 +182,5 @@ public class EaseeChargerHandler extends BaseThingHandler implements EaseeThingH
                 throw new UnsupportedOperationException(
                         "write command not found for channel: " + channel.getUID().getIdWithoutGroup());
         }
-    }
-
-    @Override
-    public Logger getLogger() {
-        return logger;
     }
 }
